requirements.txt
aiogram==3.19.0
asposestorage==1.0.2
lightrag_hku==1.3.0
transformers==4.51.0
scipy==1.14.1
numpy==2.0.2
graspologic
dotenv
lightrag-hku 
tiktoken 
pipmaster
pymupdf
python-docx